#3: Bug fix in time units got from the CmdLineUi - ideally should pass around TimeSlot objects instead of mixtures of longs and doubles. Also fix to allow less than one second idle times to be allocated properly.

+++ IdleTimer/src/idletimer/ui/TimeAllocationChooser.java
@@ -39,9 +39,9 @@ public interface TimeAllocationChooser {
 	 * Get the total amount of time the user has selected to assign to the
 	 * chosen task.
 	 * 
-	 * @return The time to allocate, in milliseconds.
+	 * @return The time to allocate, in seconds.
 	 */
-	public long GetAmountOfTimeToAllocate();
+	public double GetAmountOfTimeToAllocate();
 	
 	/** Get the task that the user chose to allocate the time to.
 	 * @return The chose task instance.
